esp_matter: Change all IDs to their correct data types.

+++ components/esp_matter/esp_matter_attribute_utils.h
@@ -197,7 +197,7 @@ typedef enum callback_type {
  * @return ESP_OK on success.
  * @return error in case of failure.
  */
-typedef esp_err_t (*callback_t)(callback_type_t type, int endpoint_id, int cluster_id, int attribute_id,
+typedef esp_err_t (*callback_t)(callback_type_t type, uint16_t endpoint_id, uint32_t cluster_id, uint32_t attribute_id,
                                 esp_matter_attr_val_t *val, void *priv_data);
 
 /** Set attribute callback
@@ -228,7 +228,7 @@ esp_err_t set_callback(callback_t callback, void *priv_data);
  * @return ESP_OK on success.
  * @return error in case of failure.
  */
-esp_err_t update(int endpoint_id, int cluster_id, int attribute_id, esp_matter_attr_val_t *val);
+esp_err_t update(uint16_t endpoint_id, uint32_t cluster_id, uint32_t attribute_id, esp_matter_attr_val_t *val);
 
 /** Attribute value print
  *
@@ -239,7 +239,7 @@ esp_err_t update(int endpoint_id, int cluster_id, int attribute_id, esp_matter_a
  * @param[in] attribute_id Attribute ID of the attribute.
  * @param[in] val Pointer to `esp_matter_attr_val_t`. Appropriate elements should be used as per the value type.
  */
-void val_print(int endpoint_id, int cluster_id, int attribute_id, esp_matter_attr_val_t *val);
+void val_print(uint16_t endpoint_id, uint32_t cluster_id, uint32_t attribute_id, esp_matter_attr_val_t *val);
 
 } /* attribute */
 } /* esp_matter */
