+++ examples/zap_light/main/app_driver.cpp
@@ -20,7 +20,7 @@
 using namespace esp_matter;
 
 static const char *TAG = "app_driver";
-extern int light_endpoint_id;
+extern uint16_t light_endpoint_id;
 
 /* Do any conversions/remapping for the actual value here */
 static esp_err_t app_driver_light_set_power(esp_matter_attr_val_t *val)
@@ -49,9 +49,9 @@ static esp_err_t app_driver_light_set_saturation(esp_matter_attr_val_t *val)
 static void app_driver_button_toggle_cb(void *arg)
 {
     ESP_LOGI(TAG, "Toggle button pressed");
-    int endpoint_id = light_endpoint_id;
-    int cluster_id = OnOff::Id;
-    int attribute_id = OnOff::Attributes::OnOff::Id;
+    uint16_t endpoint_id = light_endpoint_id;
+    uint32_t cluster_id = OnOff::Id;
+    uint32_t attribute_id = OnOff::Attributes::OnOff::Id;
 
     uint8_t value;
     attribute::get_val_raw(endpoint_id, cluster_id, attribute_id, &value, sizeof(uint8_t));
@@ -60,7 +60,8 @@ static void app_driver_button_toggle_cb(void *arg)
     attribute::update(endpoint_id, cluster_id, attribute_id, &val);
 }
 
-esp_err_t app_driver_attribute_update(int endpoint_id, int cluster_id, int attribute_id, esp_matter_attr_val_t *val)
+esp_err_t app_driver_attribute_update(uint16_t endpoint_id, uint32_t cluster_id, uint32_t attribute_id,
+                                      esp_matter_attr_val_t *val)
 {
     esp_err_t err = ESP_OK;
     if (endpoint_id == light_endpoint_id) {
@@ -90,9 +91,9 @@ esp_err_t app_driver_attribute_set_defaults()
     /* Get the default value (current value) from matter submodule and update the app_driver */
     esp_err_t err = ESP_OK;
     uint8_t value;
-    int endpoint_id = 0;
-    int cluster_id = 0;
-    int attribute_id = 0;
+    uint16_t endpoint_id = 0;
+    uint32_t cluster_id = 0;
+    uint32_t attribute_id = 0;
     esp_matter_attr_val_t val = esp_matter_invalid(NULL);
 
     endpoint_id = light_endpoint_id;
